Replace microlens with lens-simple
As of 2016-07-09, microlens-0.4.5.0 has broken Haddocks with GHC 8.
I believe this is due to a Haddock bug, not a microlens bug.
Nonetheless the problem remains, and it prevents all Haddocks from
being built unless the Stack --no-haddock-deps flag is used, which
makes reading Haddocks much harder.  So I am switching to
lens-simple, which does not have this problem.