Change from lens-simple to lens
It seems lens-simple is not being updated anymore.  Previously I
hesitated to use lens because it has a huge dependency graph.  However,
lens seems to be well-maintained, and it is widely used so it will
probably continue to be well-maintained.  With stack and stackage, the
big dependency graph is less of an issue, as stack shares dependencies
to minimize build time and because stackage helps ensure all the
dependencies remain buildable and that you'll get a build plan (unlike
the old horror days of cabal-install and Hackage.)